+++ pkg/config/openapitests/openapi_factory_test.go
@@ -1,7 +1,8 @@
-package config
+package openapitests
 
 import (
 	"encoding/json"
+	. "github.com/portpowered/infinite-you/pkg/config"
 	"strings"
 	"testing"
