+++ pkg/config/runtime_config.go
@@ -5,6 +5,7 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+	"time"
 
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 )
@@ -518,4 +519,102 @@ func mergeStringMap(base, runtime map[string]string) map[string]string {
 		merged[key] = value
 	}
 	return merged
+}
+
+func loadRuntimeDefinitionLookupMapsFromFactoryConfig(factoryDir string, cfg *interfaces.FactoryConfig, opts InlineRuntimeDefinitionOptions) (*runtimeDefinitionLookupMaps, error) {
+	if cfg == nil {
+		return newRuntimeDefinitionLookupMaps(0, 0), nil
+	}
+
+	runtimeDefs := newRuntimeDefinitionLookupMaps(len(cfg.Workers), len(cfg.Workstations))
+
+	for _, workstation := range cfg.Workstations {
+		def, err := runtimeWorkstationDefinition(factoryDir, workstation, opts.RequireSplitDefinitions, opts.WorkstationLoader)
+		if err != nil {
+			return nil, fmt.Errorf("load workstation %q config: %w", workstation.Name, err)
+		}
+		if def != nil {
+			runtimeDefs.workstations[workstation.Name] = def
+		}
+	}
+
+	for _, worker := range cfg.Workers {
+		def, err := runtimeWorkerDefinition(factoryDir, worker, opts.RequireSplitDefinitions)
+		if err != nil {
+			return nil, fmt.Errorf("load worker %q config: %w", worker.Name, err)
+		}
+		if def != nil {
+			runtimeDefs.workers[worker.Name] = def
+		}
+	}
+
+	return runtimeDefs, nil
+}
+
+// applyRuntimeDefinitionsToClonedFactoryConfig mutates a cloned factory config
+// in place so runtime worker and workstation definitions follow one merge path.
+func applyRuntimeDefinitionsToClonedFactoryConfig(cfg *interfaces.FactoryConfig, runtimeCfg interfaces.RuntimeDefinitionLookup) error {
+	if cfg == nil {
+		return nil
+	}
+
+	for i := range cfg.Workers {
+		if runtimeCfg == nil {
+			continue
+		}
+		def, ok := runtimeCfg.Worker(cfg.Workers[i].Name)
+		if !ok || def == nil {
+			continue
+		}
+		applyWorkerRuntimeDefinition(&cfg.Workers[i], def)
+	}
+
+	for i := range cfg.Workstations {
+		normalizeCanonicalWorkstationRuntime(&cfg.Workstations[i])
+		if runtimeCfg == nil {
+			continue
+		}
+		def, ok := runtimeCfg.Workstation(cfg.Workstations[i].Name)
+		if !ok || def == nil {
+			continue
+		}
+		if err := applyWorkstationRuntimeDefinition(&cfg.Workstations[i], def); err != nil {
+			return fmt.Errorf("normalize workstation %q config: %w", cfg.Workstations[i].Name, err)
+		}
+	}
+
+	return nil
+}
+
+// NormalizeWorkstationExecutionLimit rewrites legacy workstation timeout
+// authoring into the canonical limits.maxExecutionTime field and clears the
+// retired top-level timeout field.
+func NormalizeWorkstationExecutionLimit(cfg *interfaces.FactoryWorkstationConfig) {
+	if cfg == nil {
+		return
+	}
+	if strings.TrimSpace(cfg.Limits.MaxExecutionTime) == "" && strings.TrimSpace(cfg.Timeout) != "" {
+		cfg.Limits.MaxExecutionTime = cfg.Timeout
+	}
+	cfg.Timeout = ""
+}
+
+// WorkstationExecutionTimeout resolves the configured workstation execution
+// timeout from the canonical execution limits field.
+func WorkstationExecutionTimeout(cfg *interfaces.FactoryWorkstationConfig) (time.Duration, error) {
+	if cfg == nil {
+		return 0, nil
+	}
+
+	if strings.TrimSpace(cfg.Limits.MaxExecutionTime) != "" {
+		timeout, err := time.ParseDuration(cfg.Limits.MaxExecutionTime)
+		if err != nil {
+			return 0, fmt.Errorf("invalid workstation limits.maxExecutionTime %q: %w", cfg.Limits.MaxExecutionTime, err)
+		}
+		if timeout > 0 {
+			return timeout, nil
+		}
+	}
+
+	return 0, nil
 }
